feat: add GitHub and Google login buttons; refactor login component to use new button components
